requirements.txt
mcp==0.9.0
anthropic==0.39.0
langchain==0.1.20
langchain-anthropic==0.1.11
langchain-community==0.0.38
faiss-cpu==1.7.4
sentence-transformers==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
supabase==2.4.0
